move coloredtestrunner.py into tests/
