+++ test.py
@@ -13,7 +13,7 @@ import os
 import sys
 import traceback
 import unittest2 as unittest
-from coloredtestrunner import ColoredTestRunner, ColoredTestResult
+from tests.coloredtestrunner import ColoredTestRunner, ColoredTestResult
 
 flags = {'pynwb': 2, 'integration': 3, 'example': 4}
